add credit

+++ src/main/java/com/bank/models/User.java
@@ -1,10 +1,9 @@
 package com.bank.models;
 
 import com.bank.enums.Role;
-import java.math.BigDecimal;
 
 public class User {
-    private int customerId;
+    private Long customerId;
     private String name;
     private String email;
     private String phoneNumber;
@@ -13,11 +12,9 @@ public class User {
     private Role role;
     private String password;
 
-
-
     // Getters & setters (password now has a setter)
-    public Integer getCustomerId() { return customerId; }
-    public void setCustomerId(Integer customerId) { this.customerId = customerId; }
+    public Long getCustomerId() { return customerId; }
+    public void setCustomerId(Long customerId) { this.customerId = customerId; }
 
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
